requirements.txt
asgiref==3.7.2
boto3==1.26.146
botocore==1.29.146
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.16
